requirements.txt
asgiref==3.8.1
astroid==3.2.4
attrs==24.2.0
colorama==0.4.6
dill==0.3.8
Django==4.2
django-filter==23.2
django-seed==0.3.1
django-tenants==3.6.1
djangorestframework==3.15.2
Faker==26.3.0
isort==5.13.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
mccabe==0.7.0
pip==24.2
platformdirs==4.2.2
psycopg2-binary==2.9.9
pylint==3.2.6
python-dateutil==2.9.0.post0
pytz==2024.1
referencing==0.35.1
rpds-py==0.20.0
six==1.16.0
sqlparse==0.5.1
tomlkit==0.13.2
toposort==1.10
tzdata==2024.1
